KEYS: Don't permit request_key() to construct a new keyring
If request_key() is used to find a keyring, only do the search part - don't
do the construction part if the keyring was not found by the search.  We
don't really want keyrings in the negative instantiated state since the
rejected/negative instantiation error value in the payload is unioned with
keyring metadata.

Now the kernel gives an error:

	request_key("keyring", "#selinux,bdekeyring", "keyring", KEY_SPEC_USER_SESSION_KEYRING) = -1 EPERM (Operation not permitted)

Change-Id: I78657a52270c43f71c0a80af1310b39c35344ec3
Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>